PaperHub/requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
pydantic
sentence-transformers
arxiv
python-dotenv
alembic
pandas
numpy
torch
requests